RavenDB-25651 - Addressed PR Comments
