+++ test/SlowTests/Server/Documents/AI/ChatCompletionClientTests.cs
@@ -9,6 +9,7 @@ using FastTests;
 using Newtonsoft.Json;
 using Raven.Client.Documents.Operations.AI;
 using Raven.Server.Documents.AI;
+using Raven.Server.Documents.AI.Settings;
 using Raven.Server.Logging;
 using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
@@ -63,20 +64,17 @@ public class ChatCompletionClientTests : RavenTestBase
                 "{\"Text\":\"Surefire investment property in caiman islands, win $$$$ for sure, qucik!\",\"Author\":\"homepage\",\"Id\":\"2236672c-b941-4855-999e-5374f41cbddd\"}";
 
             (string Result, string Message) res = (null, null);
-            bool succeeded;
             try
             {
                 res = await client.TestCompleteAsync(prompt, context, defaultJsonSchema, default);
                 var answer = JsonConvert.DeserializeObject<AiCommentResult>(res.Result); // check if it can be parsed to json, if cannot parse it throws
                 Assert.NotNull(answer.Blocked);
                 Assert.False(string.IsNullOrEmpty(answer.Reason));
-                succeeded = true;
             }
             catch (RefusedToAnswerException)
             {
-                succeeded = true; // expected - the llm can refuse answering this because it's a violent prompt
+                // expected - the llm can refuse answering this because it's a violent prompt
             }
-            Assert.True(succeeded);
         }
     }
 
@@ -142,14 +140,14 @@ public class ChatCompletionClientTests : RavenTestBase
             using (var stream = new MemoryStream(Encoding.UTF8.GetBytes(json)))
             {
                 var blt = await context.ReadForMemoryAsync(stream, "json");
-                Assert.True(ChatCompletionClient.AiResponseParser.GetFiltersMessage(blt, out var refusal));
+                Assert.True(AzureOpenAiChatCompletionClientSettings.GetFiltersMessage(blt, out var refusal));
                 Assert.Equal("Response blocked due to content policy: hate (high severity), protected_material_code (detected severity), violence (medium severity)", refusal);
             }
 
             using (var stream = new MemoryStream(Encoding.UTF8.GetBytes(json2)))
             {
                 var blt = await context.ReadForMemoryAsync(stream, "json2");
-                Assert.False(ChatCompletionClient.AiResponseParser.GetFiltersMessage(blt, out var refusal));
+                Assert.False(AzureOpenAiChatCompletionClientSettings.GetFiltersMessage(blt, out var refusal));
                 Assert.Empty(refusal);
             }
         }
